# services/backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Personal Fitness Planner API
  summary: Cloud API for synchronized personal training plans and workout records
  description: Server-authoritative exercise catalog and immutable training plan versions, with offline-friendly idempotent
    workout synchronization.
  version: 1.0.0
paths:
...
  /api/v1/equipment:
    get:
      tags:
      - catalog
      summary: List Equipment
      operationId: list_equipment_api_v1_equipment_get
      security:
      - HTTPBearer: []
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 200
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentPage'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/sync/changes:
    get:
      tags:
      - synchronization
      summary: Get Sync Changes
      operationId: get_sync_changes_api_v1_sync_changes_get
      security:
      - HTTPBearer: []
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 200
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncChangesOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sync/batch:
    post:
      tags:
      - synchronization
      summary: Sync Batch
      operationId: sync_batch_api_v1_sync_batch_post
      security:
      - HTTPBearer: []
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 128
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncBatchIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncBatchOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    EquipmentOut:
      properties:
        id:
          type: string
          title: Id
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        category:
          type: string
          title: Category
        brand:
          anyOf:
          - type: string
          - type: 'null'
          title: Brand
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
      type: object
      required:
      - id
      - code
      - name
      - category
      - version
      - created_at
      - updated_at
      title: EquipmentOut
    EquipmentPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/EquipmentOut'
          type: array
          title: Items
        cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
        has_more:
          type: boolean
          title: Has More
          default: false
      type: object
      title: EquipmentPage
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SyncBatchIn:
      properties:
        batch_id:
          anyOf:
          - type: string
            format: uuid
          - type: string
          title: Batch Id
        sent_at:
          type: string
          format: date-time
          title: Sent At
        operations:
          items:
            $ref: '#/components/schemas/SyncOperationIn'
          type: array
          maxItems: 100
          minItems: 1
          title: Operations
      type: object
      required:
      - batch_id
      - sent_at
      - operations
      title: SyncBatchIn
    SyncBatchItemResult:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        client_outbox_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Client Outbox Id
        status:
          type: string
          title: Status
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        server_version:
          anyOf:
          - type: integer
          - type: 'null'
          title: Server Version
        server_copy:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Server Copy
      type: object
      required:
      - id
      - status
      title: SyncBatchItemResult
    SyncBatchOut:
      properties:
        batch_id:
          anyOf:
          - type: string
            format: uuid
          - type: string
          - type: 'null'
          title: Batch Id
        results:
          items:
            $ref: '#/components/schemas/SyncBatchItemResult'
          type: array
          title: Results
        accepted_outbox_ids:
          items:
            type: string
            format: uuid
          type: array
          title: Accepted Outbox Ids
        cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      type: object
      title: SyncBatchOut
    SyncChangeOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        entity_type:
          type: string
          title: Entity Type
        entity_id:
          type: string
          format: uuid
          title: Entity Id
        operation:
          type: string
          title: Operation
          default: UPSERT
        version:
          type: integer
          title: Version
          default: 1
        payload:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Payload
        changed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Changed At
      type: object
      required:
      - id
      - entity_type
      - entity_id
      title: SyncChangeOut
    SyncChangesOut:
      properties:
        changes:
          items:
            $ref: '#/components/schemas/SyncChangeOut'
          type: array
          title: Changes
        cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
        has_more:
          type: boolean
          title: Has More
          default: false
        full_resync_required:
          type: boolean
          title: Full Resync Required
          default: false
      type: object
      title: SyncChangesOut
    SyncOperationIn:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        client_outbox_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Client Outbox Id
        idempotency_key:
          type: string
          maxLength: 128
          minLength: 1
          title: Idempotency Key
        entity_type:
          type: string
          maxLength: 64
          minLength: 1
          title: Entity Type
        entity_id:
          type: string
          format: uuid
          title: Entity Id
        operation:
          type: string
          maxLength: 32
          minLength: 1
          title: Operation
        payload:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Payload
      type: object
      required:
      - id
      - idempotency_key
      - entity_type
      - entity_id
      - operation
      title: SyncOperationIn
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer